Allow 'save_study' header keyword to be None
This header variables was none in old version of AMMR (2.2), and we currently use the new version of AnyPyTools for all old AMMR tests. Hence we need this to be backwards compatible. This should hopefull not be needed in the future, when we can use pixi and run old test with old versions of AnyPyTools.